+++ src/test/java/org/opengis/cite/geomatics/VerifyTopologicalRelationships.java
@@ -36,8 +36,9 @@ public class VerifyTopologicalRelationships {
 				"/gml/Point-UTM32N.xml"));
 		Document polygon = docBuilder.parse(this.getClass()
 				.getResourceAsStream("/gml/Polygon.xml"));
-		boolean intersects = TopologicalRelationships.intersects(
-				point.getDocumentElement(), polygon.getDocumentElement());
+		boolean intersects = TopologicalRelationships.isSpatiallyRelated(
+				"Intersects", point.getDocumentElement(),
+				polygon.getDocumentElement());
 		Assert.assertTrue("Expected point to intersect polygon.", intersects);
 	}
 
@@ -48,8 +49,9 @@ public class VerifyTopologicalRelationships {
 				"/gml/LineString.xml"));
 		Document polygon = docBuilder.parse(this.getClass()
 				.getResourceAsStream("/gml/Polygon.xml"));
-		boolean intersects = TopologicalRelationships.intersects(
-				curve.getDocumentElement(), polygon.getDocumentElement());
+		boolean intersects = TopologicalRelationships.isSpatiallyRelated(
+				"Intersects", curve.getDocumentElement(),
+				polygon.getDocumentElement());
 		Assert.assertFalse("Expected curve and polygon to be disjoint.",
 				intersects);
 	}
@@ -61,8 +63,9 @@ public class VerifyTopologicalRelationships {
 				.getResourceAsStream("/gml/MultiSurface.xml"));
 		Document point = docBuilder.parse(this.getClass().getResourceAsStream(
 				"/gml/Point.xml"));
-		boolean intersects = TopologicalRelationships.intersects(
-				mSurface.getDocumentElement(), point.getDocumentElement());
+		boolean intersects = TopologicalRelationships.isSpatiallyRelated(
+				"Intersects", mSurface.getDocumentElement(),
+				point.getDocumentElement());
 		Assert.assertTrue("Expected Point and MultiSurface to intersect.",
 				intersects);
 	}
@@ -74,8 +77,9 @@ public class VerifyTopologicalRelationships {
 				.getResourceAsStream("/gml/MultiSurface.xml"));
 		Document point = docBuilder.parse(this.getClass().getResourceAsStream(
 				"/gml/Point-srsName-http.xml"));
-		boolean intersects = TopologicalRelationships.intersects(
-				mSurface.getDocumentElement(), point.getDocumentElement());
+		boolean intersects = TopologicalRelationships.isSpatiallyRelated(
+				"Intersects", mSurface.getDocumentElement(),
+				point.getDocumentElement());
 		Assert.assertTrue("Expected Point and MultiSurface to intersect.",
 				intersects);
 	}
@@ -89,8 +93,9 @@ public class VerifyTopologicalRelationships {
 				"/gml/Point-UTM10N.xml"));
 		Document envelope = docBuilder.parse(this.getClass()
 				.getResourceAsStream("/gml/Envelope.xml"));
-		boolean intersects = TopologicalRelationships.intersects(
-				point.getDocumentElement(), envelope.getDocumentElement());
+		boolean intersects = TopologicalRelationships.isSpatiallyRelated(
+				"Intersects", point.getDocumentElement(),
+				envelope.getDocumentElement());
 		Assert.assertFalse(intersects);
 	}
 
@@ -101,8 +106,9 @@ public class VerifyTopologicalRelationships {
 				.getResourceAsStream("/gml/MultiCurve.xml"));
 		Document line2 = docBuilder.parse(this.getClass().getResourceAsStream(
 				"/gml/LineString-2.xml"));
-		boolean intersects = TopologicalRelationships.intersects(
-				multiCurve.getDocumentElement(), line2.getDocumentElement());
+		boolean intersects = TopologicalRelationships.isSpatiallyRelated(
+				"Intersects", multiCurve.getDocumentElement(),
+				line2.getDocumentElement());
 		Assert.assertTrue(
 				"Expected MultiCurve and LineString (UTM) to intersect.",
 				intersects);
@@ -115,8 +121,9 @@ public class VerifyTopologicalRelationships {
 				"/gml/Curve-LineString.xml"));
 		Document curve2 = docBuilder.parse(this.getClass().getResourceAsStream(
 				"/gml/LineString.xml"));
-		boolean disjoint = TopologicalRelationships.disjoint(
-				curve1.getDocumentElement(), curve2.getDocumentElement());
+		boolean disjoint = TopologicalRelationships.isSpatiallyRelated(
+				"Disjoint", curve1.getDocumentElement(),
+				curve2.getDocumentElement());
 		Assert.assertTrue("Expected curves to be disjoint.", disjoint);
 	}
 }